Fix #243: weak pointer to static array
Add function `clear-static-vector-mark` to clear the static vector
mark for all the static vectors in `lisp::*static-vectors*`.  This is
added to `*before-gc-hooks*` so that the marks are cleared before GC
so that GC can set it if the static vector is still in use.

Modify `finalize-static-vectors` to delete any element in
`lisp::*static-vectors*` which has broken weak pointer.  That means GC
has freed the static vectors so we don't need to keep track of it
anymore.

Finally, modify `scan_weak_pointers` to free the static vectors.  As
the list of weak pointers is scanned, find any weak pointers to static
arrays.  If the mark on the static array is clear, add it to another
list if it's not already there.  At this point we can also break the
weak pointer after it's added.  After looking at all the weak
pointers, we can look at this list of static arrays and free them.
Since the list is unique, we don't double-free anything.