Fix some logic errors.
When checking the header value, we want a header value of exactly 1,
which means that we have a static vector and that the mark is clear.
This means the static vector is not referenced by anything.

We were also breaking the weak pointer in the wrong place.  We only
want to break the weak pointer when the static vector can be freed.
Otherwise, leave the weak pointer alone.