Add function to add to free list
Make the `scan_weak_pointers` a bit smaller by moving the code for
inserting a value to the free list into its own function.

Update some comments and fix a printf.  The printf was printing the
address of the header instead of the value, which is what we wanted to
so we can see that we do have a static array and whether it was marked
as live or not.