Use integer-length instead of log2
For positive integers, `integer-length` is close enough to `log2` for
what we want; we only need an approximate value for log2(x) to know if
x is too small or too large to fit in a float.  This should make
things a bit faster so this change has less of an impact on the speed
of reading floats.