Create list of all static vectors, not just the unmarked ones
Instead of creating a list of unmarked static vectors, create a list
of all static vectors.  Then we can process the unmarked vectors and
free them carefully.

Although we haven't implemented it yet, this allows us to make one
final pass to unmark the marked static vectors so that when GC happens
again, we can mark them as being in use.