Clean up implementation
Add `#defines` for the various bits needed to handle processing of
static arrays.

When scavenging weak pointers to static arrays, clear out the visited
bit, just to be safe.

Clean up debugging prints.

Add phase 4 to go through all the remaining static vectors and unmark
them so we don't have to do that in Lisp.