Be more careful about clearing the static vector mark bit
Check that the weak pointer hasn't been broken yet before trying to
clear the mark bit of the static vector.

I think this currently doesn't matter because when the weak pointer is
broken, the value is set to NIL, and NIL doesn't have the MSB set.
But that's not guaranteed, so let's be safe.