Remove broken weak pointers from the list
In phase 2 where we break weak pointers that point to a static vector
that we've already visited, we can delete this broken weak pointer
from our list.  We don't need to do anything more with this weak
pointer.

Likewise in phase 3, when we break the weak pointer to unused static
vector, we can also remove this weak pointer from the list since
there's nothing more we need to do.

In phase 3 and phase 4, add an assert that the vector we're examining
is not broken.  They should have been removed in the previous phase.