Remove visited weak pointer to static vector in phase 2
In phase 2, if we've previously visited a static vector, we break the
weak pointer to it.  In this change, we also destructively remove the
weak pointer from `static_vector_list` so that the following phase
doesn't have any broken weak pointers.

In destructively removing the weak pointer, we have take care of the
case where the weak pointer is the first one.  We also need to handle
the case where the weak pointer is the last in the list.

In phase 3, add an assert that the weak pointer has not been broken.

The two test cases we have work as expected.