Remove unneeded pointers during processing.
Add `pop` and `push` functions to handle the lists we need to keep
track of things.

In phase 2, live vectors are added to the in-use list.  For unused
vectors, the first visit adds them to the freeable list.  Subsequent
visits to this static vector break the weak pointer.

In phase 3, process the freeable list by freeing the vector and breaking
the corresponding weak pointer to the vector.

In phase 4, process the in-use list and unmark the vectors so the next
GC can tell if the vector is in use or not.