Unmark static vectors at the end of GC
Before, we were unmarking live static vectors at the end of
`scan_static_vectors`.  But apparently this is not quite right because
when doing `(gc :full t)` multiple times, we clear the mark bit to
static vectors that are actually still alive and that, of course,
causes havoc.

Instead, unmark the vectors at the end of collect_garbage when we've
finished all GCs.  This kind of matches how we were originally
unmarking the static vectors in a post-GC hook.

With this change when running our weak-pointer test, `(gc :full t)`
can be called many times without causing issues.