Clear static vector visited bit when scavenging a weak pointer to it
When scavenging the weak pointer to a static vector, clear out the
visited bit to prevent confusion when scanning all the weak pointers.
This is needed because scanning causes use to unmark live vectors and
set the visited bit so we don't add other weak pointers to the now
unmarked (free) static vector.

Undo the change in `scav_static_vector` so we just the mark bit and
don't clear the visited bit.  That whould happen when we scavenge the
weak pointer.