Fix #293:  Allow restarts for FP overflow in reader
When a FP overflow occurs when reading a float, allow the user to
restart the computation by returning a floating-point infinity or by
returning the largest float.

The function `restart-overflow` performs the restart and returns the
desired value if so selected.

`make-float-aux` calls `restart-overflow` when the number is known to
be too large or when a floating-point-overflow is signaled from
`coerce`.

However to make that work, `check-exponent` in `bignum-to-float-float`
must signal a `floating-point-overflow` instead of
`simple-type-error`.