Fix #299: Enable xoroshiro assembly routine
Use the assembly routine (via the generated VOP named
`kernel::random-xoroshiro-update` in `kernel::xoroshiro-gen`.

This allows us to remove the VOP `vm::xoroshiro-gen` from arith.lisp
too.

Manually built and tested this using a build from !202.  This works
and the unit tests and ansi tests pass.

Can't currently CI with this, so [SKIP-CI].