Fix the retry limit for resizing password entry buffer
The previous version checked that the buffer had not exceeded the
limit of 1MiB after the buffer was resized.  In otherwords, the buffer
is allowed to grow past the limit before the retry loop is exited.

This change moves the check of the buffer size before the allocation
occurs.  Now, the retry loop is exited before the buffer would grow
past the limit.