Remove the definition of __NO_CTYPE from the Linux build
The definition of this macro disables the definition of isalpha(3) and
other symbols from ctype.h from being defined as macros that depend on
internal glibc functions not present in all versions of glibc.

At the time, that allowed binaries compiled against newer versions of
glibc to run on systems with older versions of glibc.  However, today
the internal functions have been available for more than two decades
and so this feature does not appreciably affect compatibility.