Fix #316:  Make (code-char 181) have no case
For compatibility with other lisps
where `(char-equal (code-char
c) (char-downcase (char-upcase (code-char c)))` is true for all `c`
below 256, adjust cmucl to do the same.

This is done by change the limit where we use Unicode rules for the
character functions.  Previously, this was 127 but we are changing it
to 191.  Added a new constant `+unicode-lower-limit+` for this value
so that we can consistently change it for all functions that need it.