Adjust both-case-p, lisp::unicode-upper/lower
For `both-case-p`, Unicode characters have both cases only if the
Unicode category is upper or lower.  Previously title case characters
were included.

For lisp::unicode-upper, we don't want to convert convert to upper
case if the Unicode category is Lt, Mn, Nl, or So.  These characters
won't roundtrip a case conversion.

Likewise, for lisp::unicode-lower, we don't convert to lower case if
the category is Lt, Nl, or So.

Also change the unicode lower limit back to 127.

With these changes, The ansi-tests char-upcase.2, char-downcase.2, and
both-case-p.2 now pass.