Slightly simplify implementation of char-upcase/downcase transforms
First, change the arg type from `base-char` to `character` (which are
the same in cmucl).

Second, for `char-upcase`, we only upcase a Unicode character if it is
a lower-case letter (category "Ll").  Likewise for `char-downcase`, we
only downcase if the character is an upper-case letter (category
"Lu").  Everything else is ignored.