Add comments and clean up code.
Document the functions and clean up implementation, renaming some and
adding more.

Add `find-optimum-size` function to try all stage2 sizes to find the
one that minimizes the space needed for the tables.  We ran this to
find out the optimum size is to have a stage2 size of 6.

Add 'build-case-table` to build the case table and print the results
to a file, defaulting to "src/lisp/case-table.c".