Implement suggested modifications
First, when a second stage table is all zeroes, don't use NULL, but
use the same array of all zeroes.  Then we don't have to check if the
second stage array is empty.  It can be used like any other array.

Second, change the contents of the table back to containing a delta
instead of the actual code.  This simplifies computing the code
because we don't need to check for zeroes.  Just always apply the
delta.

Third, update case-table.c using the new generator code.