Handle NaN in ir1-transform->=
Previously, when `x` and `y` are the same leaf, the transform would
immediately return `T`.  However, this is not correct if `x` is a
float because NaN >= NaN is always false.  So add check that `x` is
not a float.

Manually tested (for now) the with
```
(defun bar (x)
  (declare (double-float x))
  (<= x y))
```
returns `NIL` for `(bar *nan*)` and `T` for `(bar pi)`.  Without this,
it always returned `T`.

Update cmucl.pot too because it was regenerated.