Update ansi-test on the correct branch.
Previously, if the ansi-test directory existed, we did a `git pull
--rebase` on whatever branch we happened to be on.  Then a bit later
we switched to the desired branch.  This meant that we updated some
branch, but then didn't get the updates for the branch we actually
wanted to sue.

Modify script so that we change branches to the desired branch before
pulling new changes.

Since the branch name is used in a couple of places, add an envvar to
hold the name of the desired branch.