More cleanups
Just use simple `(string string)` instead of doing a test for
`stringp`.

Fix bad formatting in `string-upcase-full`.

Use `ecase` as needed and also add declarations to it clearer to the
compiler what the acceptable arguments are.

Also, we were confused on what `string-capitalize-simple` does compared
to `string-capitalize-full`.  We call `string-capitalize-simple` if
the `casing` arg is `:simple-title` instead of `:simple`, which should
call `string-capitalize-full`.