Fix #333: LOAD :IF-DOES-NOT-EXIST is a generalized boolean
First, we need to update the compiler's function database with the
correct type for `:if-does-not-exist`.  The value can be anything, not
just a `(member :error nil)`.

Second, in `load`, convert any non-NIL value of `:if-does-not-exist`
to `:error` before calling `internal-load` which expect it to be
either `:error` or `nil`.  (It's passed do `open`).

Finally, add a couple of simple tests for this.