Fix note about unreachable code in float-sse2.lisp
The warning happens because we do `(and ,commutative ...)` where
`commutative` is a known value passed to the macro.  When the value is
zero, we get the note since the rest of the `and` clause is known to
be unreachable.

Change this to handle the `commutative` and macro-expansion time to
add the clause only when `commutative` is true.