Fix #337:  Move +ascii-limit+ to srctran.lisp
When cross-compiling from linux/x86 to linux/x86, we cross-compile
srctran.lisp.  It appears that this fails because the constant `+ascii-limit+` is
not available because it's defined in char.lisp.

A solution is to define this constant in srctran.lisp (so the
cross-compiler has it) and update all uses of `+ascii-limit+` to use
the value from `C` package in srctran.lisp.