Address #336: get-nil-indexed-object returns object and validp value
To fix the issue where compiling `x86::print-imm-data` printed a
compiler note about unreachable code, we modify what
`get-nil-indexed-object` does.  This now returns two values:  the
lisp object and a validp flag which is true if the offset fits in the
static space, which has all of the possible nil-indexed objects.

Update `maybe-not-nil-indexed-object` to handle the extra return
values.  If the second value is not true, we skip printing a note.
This makes the comment about returning non-NIL iff a note was recorded
a true statement.