Handle leak of searched_core
The initial fix of just returning NULL when `core` is NULL or when
`core` is not accessible like so:
```
     if (core && access(core, R_OK) != 0) {
       return NULL;
     }
```
fixed the analyzer warning that we were leaking `core`.

I think the analyzer is wrong here.  If `core` was set by
`search_core` but it is inaccessible, we'd return NULL without freeing
the space returned by search_core.

Instead we do this so that we the searched core is NULL or if it's not
accessible, we free the space before returning.  This also fixes the
analyzer warning.