Fix #262:  arch_skip_inst invalid code
`arch_skip_instruction` was not correctly updated to handle the new
UD1 instruction for trapping.  It was just using the 3rd byte of the
instruction to get the trap code.  However, that's incorrect.  The low
6 bits of the byte represent the trap code.

Add a simple macro to extract this value and use it also in
`sigill_handler` (which was extracting the code correctly).