Fix #352:  New compression methods for distribution tarballs
Allow user to specify different compression options.  The new `-C`
option for `make-dist.sh` allows the user to choose how to compress
the tarballs for a distribution.  The default compression option is
determined from the OS but is usually xz compression.

We also update the scripts to call `tar` with the appropriate option
to compress the tar archive.  This is achieved by passing in envvars
to the support scripts.