Some cleanups and fixes.
Export `GTAR` so that make-main-dist.sh and make-extra-dist.sh use the
right version of tar.  (These scripts should probably be updated to be
consistent with make-src-dist.sh which takes an option for tar
program.)

Add usage message for the -C option.  Also change the valid values to
be `bzip2`, `xz`, and `gzip` instead of `-j`, `-J`, and `-z`.  The
names are much more informative.