Add -t option for the tar program to use
Like make-src-dist.sh, add a `-t` option to make-main-dist.sh and
make-extra-dist.sh to specify the tar program to use.  (Generally it's
`tar` except on Solaris where we want to use `gtar`.)

This means make-dist.sh has to determine the tar program to use and
sets the `-t` option when calling the other scripts.