Ignore errors when setting the modes and more cleanups
When resetting the mode bits ignore any errors that might happen if
some traps are still enabled.  Also, only update the exception part of
the mode bits.

Document what FPE_FLTSUB means.  Make a note that the fpe code info
alist only contains the relevant info for FP traps that we're
interested in handling.

With this change, an overflow and underflow are signaled as expected
and when we restart, we don't get signaled again.  But according to
the debugging prints, we do enter the sigfpe handler twice.  Divide by
zero is not handled correctly, though.

More testing needed.