Use os_set_sigcontext_fpu_modes to reset modes
x86-vm.lisp defines `(setf sigcontext-floating-point-modes)` to set
the modes in the sigcontext.  Use this in `sigfpe-handler` to reset
trap bit for the exception that was signaled.

However, there's no implementation of `os_set_sigcontext_fpu_modes`,
so implement one in solaris-os.c.

Works somewhat better than before in that we don't resignal the
exception again when we restart.  But still have problems with divide
by zero and invalid.  And we're left in a weird FP trap state where we
can't read 1d300 anymore---an overflow?