More debugging
Add C functions to get the x87 and sse2 modes from the sigcontext.
Also add functions to set the modes in the sigcontext.  Make these
available to Lisp.

Use these functions in sigfpe-handler to examine the modes and to
clear out the current exception.

Still needs work.  Overflow is signaled correctly, but after aborting,
the FP modes still contains the exception.