Fix #356:  Put x87 status word in low 16 bits of mode
`x87-floating-point-modes` combines the FPU status word and control
word into a single 32-bit value.  Currently the control word is in the
low 16-bits and the status word is in the high 16 bits.  The status
word contains the flags to show if an exception has occurred.

However, the SSE2 mscsr register has the exception flags in the low
part of the register.

For consistency, put the status word in the low 16-bits of the result
to match more closely what SSE2 does.  This simplifies debugging when
looking at the the x87 and sse2 modes independently.