Solaris/x86 needs its own with-float-traps macro
For Solaris/x86, `with-float-traps` needs to handle the x87 and sse2
mode bits separately.

The previous version used the merged value returned by
`vm:floating-point-modes`.  If the original mode had, say, the
overflow flag set, when the macro restored the modes, the overflow
flag was also set for x87 and sse2.  Later on this can cause an
expected overflow because the x87 mode has overflow set along with the
sse2 mode.  This shows up in the SIGFPE handler.

To fix this, we handle the x87 and sse2 mode bits separately.  Hence,
we don't spuriously add back the overflow bit for x87 which didn't
have it previously set.