Tighten up float-modes deftype
The existing `float-modes` type is `(unsigned-byte 24)`.  However, for
x87, the float mode can be at least 26 bits long.  This showed up
running the unit test suite.  Examining the bits in the control word
shows that 26 is the max, if you ignore the reserved bits at the high
part of the control word.

For sse2, the mxcsr register only has 16 defined bits.

Thus, split `float-modes` into separate types, one for x87 and one for sse2.