Fix #360: Add default-site-init file
Remove the old generic-site.lisp file in favor of
default-site-init.lisp, which is a modified copy of the
generic-site.lisp.

Update make-main-dist.sh to install default-site-init.lisp (instead of
generic-site.lisp.

default-site-init.lisp just sets the "target:" search-list.  The
original stuff with site names and herald-items is gone.

On startup we look for "library:site-init.lisp" first.  If that's not
found, we look for "library:default-site-init.lisp".  It is not an
error if neither of these are found.

The herald items that used to be in the init file have been moved to
the core.  So now lisp startup prints a link to the gitlab issue
tracker and another link to the support page.

Update the pot file for these changes.