unix-mkstemp returns the file descriptor and the file name
Previously, we only returned the file descriptor.  Now we return the
descriptor and the file name associated with the file.  This is
available to us because `mkstemp` modifies the template with the
actual file name.

Also, we use %name->file to encode the Lisp string into a set of
octets to pass to `mkstemp`.  This means we need to convert updated
template to a Lisp string.

Tested this using the template "/tmp/α-XXXXXX" (alpha character
followed by "-XXXXXX").  The resulting file name was "/tmp/α-L2egJE"
whith the X's appropriately replaced.