mkdtemp needs to encode/decode Lisp strings
Before calling `mkdtemp`, we need to convert the Lisp string to
octets.  On return, the octets that were passed to `mkdtemp` have been
modified to hold the directory name.  These octets need to be
converted back to a Lisp string.

Tested with the template "/tmp/α-dir-XXXXXX".  The result was
"/tmp/α-dir-jSt2SC".  I verified that this directory was actually
created.