Fix #363: Add version number to files and directories
The version number is based on the value of `git describe --dirty`.
If it looks like "snapshot-....", we remove "snapshot-" and the
remainder is the version number to use.  If it looks like a hex
release number (like 21e), we use that as is.

This requires quite a few changes.

The default cmucl lib search list is modified in lisp.c to be,
basically, "lib/cmucl/VERSION/lib".

When building, we need to add the `-lib` option so that we lisp can
find its core file because lisp has the new search list and the core
isn't there.

GNUmakefile needs updating so that we can set the correct version in
lisp.c.

The distribution scripts (make-dist, make-main-dist, make-extra-dist)
need updating to install the files in the correct directories.  This
includes installing the lisp file as lisp-<version>.

This needs more work, but building appears to work and `make-dist.sh
-SI <dir>` makes a directory structure such that
`<dir>/bin/lisp-<version>` will start with the default core file.