Fix building lisp
For build.sh, first, echo out the command used to build each stage so
that we can see that we're using the correct args.  Second, we
incorrectly used the wrong core file with the stage 3 lisp; it was
using the core from stage 2.  Third, we forgot to pass the `-lib`
option for building stage 4 so it couldn't find the core file.

For build-utils.sh, we need to add a `-lib` option so lisp can find
the core file.