Install unit tests and run-unit-tests.sh
Install all of the unit-test files to share/cmucl/version/tests.
Install run-unit-tests.sh to bin/run-unit-tests.sh

To make this work, modify run-unit-tests to add a `-d` commandline arg
that specifies the directory where the test files are located.  If no
`-d` flag is given, it defaults to how run-unit-tests.sh used to work,
which expected the tests to in "./tests/".