Fix #367:  Add method to count number of octets to encode a string
Add new method/slot to an external format to count the number of
octets needed to encode a string in the given external format.

This requires adding a new enum value `oc`.  This also means adding a
new slot, `octet-count` to `efx` structure to hold the new method.
The `define-external-format` is updated to allow specifying the new
code to implement the `octet-count` slot.

We update the external formats for :utf-8, :iso8859-1, and :ascii to
count the octets.

Finally define the function `string-octet-count` to compute the number
of octets needed to encode the string in the given external format.