Add some sanity checks for STRING-TO-OCTETS
Since `string-to-octets` can return a buffer of octets whose length is
not the same as the returned number of octets, add a check that they
are the same.  If not, then we `string-octet-count` may not be
consistent.  Also add a check that the converted number of characters
matches the length of the string.