Hack to handle composing formats in string-octet-count
I can't figure out how to make `string-octet-count` work when using
composing formats (like (:latin1 :crlf) to output CR/LF for every LF).

As a workaround, if the external format is a composing format, use
`string-to-octets` to get the number of octets.  This is not great,
but allows `string-octet-count` to work with composing formats.