Disable some tests on Darwin
I tested these on Darwin, and the templates "test-" and
""test-XXXXXXa" succeed by returning exactly these files.  Another
call would then generate an error since we can't create new files.

So it seems on Darwin, any template will work but the only time we get
a filename different from the template is if the template ends in X's.