Add -print-version option and use it in make-dist.sh to get version
make-dist.sh needs to have a version that's consistent with the
version used in lisp.  To that end, we add a new string
`cmucl_version` in lisp.c that has the version string that is used to
setup the search list.  To make this available, we also add a new
commandline option, `-print-version` that just prints this string and
exits.

In make-dist.sh get the version number from `lisp -print-version`.

In this way the created directory versions match what lisp thinks the
version is.

Without this, we sometimes end up in a situation where we can't find
the core file.  This can happen when we have a clean repo, do the
build and end up with a dirty repo because the build update some pot
files or something.  Then make-dist ends up putting everything in the
wrong directory.