Set *lisp-implementation-version* from cmucl_version
Make `lisp::*lisp-implementation-version*` have the value of the C
string `cmucl_version` that contains the actual version.

To do this, we need to make `lisp::*lisp-implementation-version*` a
static symbol.  Fortunately, there are spare static symbols so we
stole `spare-9` for this.  Currently only done for x86, but the other
OSes have spare symbols too.