Remove errno stuff from unix.lisp and move it to unix-errno.lisp
So, all the stuff related to errno in unix.lisp has been moved to
src/code/unix-errno.lisp.  This file is auto-generated in build.sh and
cross-build.sh.  The new file bin/create-errno.awk is the script that
generates the file.  It contains all of the errno stuff from
unix.lisp so that other OSes are supported.  But it generates
appropriate `def-unix-error` forms on Linux.