Remove -print-version switch in favor of -version
Change the commandline switch `-version` (and `--version`) to print
the version and exit.  Before, this was handled in lisp, but it's not
necessary now since the version is the same as
`*lisp-implementation-version*`.  This means we don't need
`-print-version` anymore.

We need to exit too, because the build scripts need to know the
version of lisp being used.

Update make-dist.sh to use `-version` instead of `-print-version`
which no longer exists.

Update cmucl.pot file for the changed docstrings.