Add :prefix keyword arg and more error checking
For `with-temporary-{file,directory}`, add a `:prefix` keyword arg to
allow the user to specify a prefix for the name of the temporary file
or directory.

Add sume error checking if `mkstemp` or `mkdtemp` fail to create the
requested object.  We just throw an error with a message with the
template and the reason (from errno).

Finally, in for the temporary stream, don't add the name of the file
to the fd-stream object.  The name, by default, ends up being
"descriptor n" where n is the fd used.