First cut at adding os_temp_path
Add os_temp_path to os-common.c to return an appropriate location
where temporary files can be placed, depending on the OS.  Currently,
MacOS, has a special method to return a secure per-user temp
directory.

For everywhere else, we look at the TMP envvar.  If it's not set, we
just use "/tmp".

Add get-os-temp-path in extensions.lisp to get the temporary path.
This is used in the macros, but get-os-temp-path doesn't yet call
`os_temp_path`.