Add configs for tests and benchmarks
There's lots of duplication for running the ansi tests, unit tests,
and benchmarks.  Add templates and configs for these so the common
stuff is in the base configs that can be used for the individual jobs.

Also fixed a couple of typos where the opensuse jobs had a tag of
"linux" instead of "opensuse".