Abstract out the common bits of the install job
All of the install jobs are basically the same except for the name of
the tarball that we want to download.  Move this to a template and add
a variable to select the name of the OS to use for the tarball name.

Implement just for linux:install for now, for testing.