Move .build_cmucl into .build_template
The script from .build_cmucl can be moved to .build_template since it's
the same for all OSes now.  Remove .build_cmucl and references to it.

Also remove commented-out calls to wget since we're using curl now.

Fix typo in linux:static-analyzer that was using the directory
"linux-4".  It's "build-4" now.