Ignore duplicate errno values
Put the awk script in the new file create-def-unix-error.awk.  This
script pretty much does what the previous version did, but it checks
to see if we have a duplicate errno value in the file and just ignores
it, assuming the first occurrence is what we want.

Update create-errno.sh to call this script.

Duplicates appear to happen on Darwin, where EOPNOTSUPP occurs twice
due to #if's selecting one or the other.  We assume here that the
latter is not what we want.

This is a gross hack.