Want the second definition for duplicates
Modify script so that if we find a duplicate, we mark the old as
something to be skipped on output.  We want the second (last)
definition.

This is needed on Darwin where we want the second definition of
EOPNOTSUPP.

Tested this with a simple C program on Darwin to print the value of
EOPNOTSUPP.  It is 102, not ENOTSUP (45).