Update for solaris/svr4 in the template.
create-errno.sh supports Solaris, so we can remove the special
definitions from the template.  This means that all the definitions
that remain are considered to be the same on any OS other than linux,
darwin, and solaris.

We left the feature as `:svr4`, but maybe it should be `:solaris`.  On
my solaris/x86 VM, both `:svr4` and `:solaris` are features.