Copy things directly from the template
Rearrangement of how create-errno.sh works.  If we don't have a list
of errno files, copy the template directly to the output since the
template contains the default errno values.

If we do, we can copy just the beginning and ending of the file,
skipping the default values.  The forms are auto-generated.  We leave
special markers in the template so we know what to copy.