Add usage message and command line options
Add -D and -S options to control how this script works.  -D says to
use the default instead of auto-generating.  -S says to copy the
result to stdout for quick viewing.

These command line options are mostly for debugging the script.

In any case, the file src/code/errno.lisp is always created.