Fix extern declarations
Updated extern declarations for message_write_int,
message_write_translation_table, and message_write_accelerator_table.

Also, in translations.c, there's the call
`message_write_int(reply,event,int_tag)` where `event` is an
`XEvent*`.  That's  not an int, so I changed it to `event->type`.  Not
sure if that's correct or not.