Fix #383:  Support build.sh -v option
The `-v` option for build.sh works again.  Specifying this sets the
cmucl version to the given string and is used to specify the directory
where lisp finds the core file and other files.

For the implementation, git-version.sh is updated to support a `-v`
option as well to do the appropriate thing for setting the value for
the file src/lisp/cmucl-version.h