Fix #349: Fix linker warning about LOAD segment with RWX permissions
When writing the elf header for the core sections, don't set the write
or execute flags for the section.  (Really, only needed to remove the
execute flag, but removing the write flag doesn't hurt either.)  This
doesn't affect the behavior of the resulting lisp executable.  It
still runs as expected.

To be sure, update CI so that we run the unit tests with the saved
exectuable.