Fix #400:  unix-resolve-links breaks on some lengths
Change how the result buffer size is increased.  Previously, it
tripled the length of the result buffer, but this can cause an error
because the resulting buffer is too short to add in the new data to
the result.

So instead, triple the size of the buffer using the fill pointer
value.

Solution proposed by Philip Fominkyh in the bug report.