Fix #401:  FILE-POSITION returns correct position after setting it
When the file position is set, we clear out any pending input to force
the next read to go to disk.  However, we forgot to reset the
`in-index`, so when we ask for the position after setting it,
we think there's pending input that hasn't been read.

Thus, reset `in-index`.