Add another file-position test
As mentioned in the comment for fd-stream-file-position, we have added
a test for the case where the input buffer doesn't have enough octets
to form a complete character.

We create a file consisting of a single character followed by 128
characters that require 4 octets for utf-8 encoding.  Hence the
in-buffer (of length 512) only has a partial utf-8 character at the
end.