First arg is the expected value in assert-equal
The file position tests fail with the last snapshot, of course, but
the error message is confusing because it says something like
"Expected 505 but saw 0".  This is because the first value of
`assert-equal` is the expected value.  With this change, the message
is now "Expected 0 but saw 505", which makes a lot more sense.