Fix SC2086 by using bash with arrays
The command line for gcc use many envvars to get the necessary
arguments and these arguments consist of many separate words that
needs to be split on spaces as usual.

https://github.com/koalaman/shellcheck/wiki/SC2086 suggests one option
is to use arrays.  We do so and of course need to switch from sh to
bash to run this script.