Pretend stage2 table are 16-bit values instead of 32-bit
Change definition of the alien stage2 array to be an array of
unsigned-short instead of unsigned-int.  (Pending change of actual
table to be uint16_t instead of uint32_t.)

Update get-{upper,lower}-case-entry to access 16 bits at a time
instead of 32.  This means adjusting the value returned by
case-mapping-offset to be a 16-bit offset instead of 32-bit offset.

ansi-tests pass.