Add bootstrap code to define UNIX package
Update the bootstrap file to define the UNIX package using the ERRNO
package.  By doing this, we don't need to unlock the UNIX package to
get these changes in.  Hence we can see unexpected changes in the UNIX
package.

Tested this by building the world with the boot-2024-08 file.  Then
used the result to rebuild the world without the bootstrap file.  No
warnings in code:exports.lisp from the UNIX package.