Implement support of package-local-nicknames
First, add a new slot to the package structure to old the local
nicknames.

Second, make `defpackage` recognize the option `:local-nicknames` and
save the alist.

Third, update `%defpackage` to accept the an additional arg for the
local nicknames.  Currently, this is an optional arg for bootstrapping
purposes.  We really want it to be a positional parameter.

Finally, update bootstrap file so we can change the package structure.